refactor: extract much logic into util file
